Refactor unit tests for a more clear structure
